[share_backup] u16 for threshold

+++ schnorr_fun/tests/share_backup.rs
@@ -7,7 +7,7 @@ use secp256kfun::{g, marker::Secret, poly, s, Scalar, G};
 fn share_backup_short() {
     let secret_poly = vec![s!(6), s!(1), s!(9)];
     let polynomial: Vec<_> = secret_poly.iter().map(|c| g!(c * G).normalize()).collect(); // some polynomial coefficients
-    let threshold = polynomial.len();
+    let threshold = polynomial.len() as u16;
     let share_index = s!(12).public();
     let secret_share = poly::scalar::eval(&secret_poly, share_index)
         .non_zero()
@@ -33,7 +33,7 @@ fn share_backup_short() {
 fn share_backup_long() {
     let secret_poly = vec![s!(1), s!(2), s!(3)];
     let polynomial: Vec<_> = secret_poly.iter().map(|c| g!(c * G).normalize()).collect(); // some polynomial coefficients
-    let threshold = polynomial.len();
+    let threshold = polynomial.len() as u16;
     let share_index = Scalar::<Secret>::from_str(
         "34f7ce653cfa8454b3463726a599ef2925736442d2d06455974d6feae9450d90",
     )
